Change to use SyntaxWalkers
The previous code was basically implementing a SyntaxWalker, but
using a foreach and a giant switch statement, which is more cleanly
represented using a SyntaxWalker.

Also switched to clearing out a HashSet that is initially populated
instead of trying to do it in 2 passes. Once the pass is complete,
the set should be empty, so we let go of all the references to all
the symbols we would otherwise be holding on to.