Add some thread safety
Previously it was possible that with a 1000 file solution, if all
1000 files were processed at the same time (which is not
inconceivable, since ProcessAsync is async), then we might
try to parse the entire solution for each file (since they all
hit the "haven't produced a model yet"). Resulting in 1,000,000
file parses.  Add a quick little lock to make sure only one
instance of the model is produced.