Skip searching the project for InternalsVisibleTo
The tool used to produce different output if run on the solution or
the individual projects within it, because we looked for the
referenced assembly from InternalsVisibleTo inside the solution,
since that meant we could be more sure of write usages.

But the difference in scenarios is not worth it, and most of the
dotnet code base doesn't actually use InternalsVisibleTo.